refactor: Change the error component to use the shared
