+++ src/components/Plans/UpdatePaymentInformation/UpdatePaymentInformation/index.js
@@ -14,7 +14,7 @@ import {
   updatePaymentInformationReducer,
   UPDATE_PAYMENT_INFORMATION_ACTIONS,
 } from '../Reducers/updatePaymentInformationReducer';
-import { UnexpectedError } from '../../PlanCalculator/UnexpectedError/index';
+import { UnexpectedError } from '../../../shared/UnexpectedError/index';
 
 const updatePaymentInformationteps = {
   paymentMethodInformation: 'payment-method-information',
